requirements.txt
<<<<<<< HEAD
click==6.7
Django==1.11.2
django-multiselectfield==0.1.7
greenlet==0.4.12
msgpack-python==0.4.8
neovim==0.1.13
olefile==0.44
Pillow==4.1.1
python-dotenv==0.6.4
pytz==2017.2
simplejson==3.10.0
=======
coverage==4.4.1
Django==1.11.2
django-multiselectfield==0.1.7
django-nose==1.4.4
mock==2.0.0
nose==1.3.7
olefile==0.44
pbr==3.1.1
Pillow==4.1.1
pytz==2017.2
simplejson==3.10.0
six==1.10.0
sure==1.4.6
yapf==0.16.2
>>>>>>> origin/testing
